requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# NLP
transformers>=4.38.0
torch>=2.1.0
sentence-transformers>=2.6.0
negspacy>=1.0.4
spacy>=3.7.0

# BERTopic / clustering
bertopic>=0.16.0
umap-learn>=0.5.6
hdbscan>=0.8.33

# LLM
anthropic>=0.25.0

# Survey-weighted statistics
statsmodels>=0.14.0

# Explainability
shap>=0.44.0

# Annotation (optional, for export)
label-studio-sdk>=0.0.32

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
